requirements.txt
albumentations==0.5.2
Pillow==8.1.0
tensorboard==2.4.1
torch==1.7.1
torchvision==0.8.2
